requirements.txt
panflute==1.10.6
PyYAML>=4.2b1
Jinja2>=2.10.1
requests==2.18.4
